game logic and word-object data handling.
In wordUtilities.js:
extractGivenWord() picks out the word string for the given word object.
extractDefinition() picks out the definition for the given word.
extractSynonyms() picks out the synonym array for the given word.
extractFrequency() picks out the number of times the word is likely to appear in any English corpus, per million words. If it does not have a frequency according to the API, returns a default value.
compareWordsMatch() creates an array of the words the user entered which are synonyms to the given word. (changed to array spreading)
compareWordsMismatch() creates an array of the words the user entered which are NOT synonyms to the given word.
createSynonymScoreObject() creates an object of each word with points according to its frequency.

In atoms.js changed name of some function calls to match wordUtilities.js.

In gameUtilities.js prepared a placeholder score calculation in calculateScoreFromFrequency(), score for a word depending on how recherché it is.